// src/types/reports/reports.ts
export interface IReportsStore {
  label: string;
  value: string;
}