// src/types/brand-wise-product/brandWiseProduct.type.ts
export interface IBrandWiseProduct {
  bandWLogoUrl: string;
  brandName: string;
  storeName: string;
  storeNameLogo: string;
  activeProduct: number;
  inActiveProduct: number;
  totalProducts: number;
  upC_GTINMissing: number;
  notReadyToSell: number;
  partiallyReadyToSell: number;
  readyToSell: number;
}